split minimum_distance_constraint.h/cc into separate files.
