+++ multibody/inverse_kinematics/BUILD.bazel
@@ -71,6 +71,8 @@ drake_cc_library(
         "gaze_target_constraint.cc",
         "kinematic_evaluator_utilities.cc",
         "minimum_distance_constraint.cc",
+        "minimum_distance_lower_bound_constraint.cc",
+        "minimum_distance_upper_bound_constraint.cc",
         "orientation_constraint.cc",
         "orientation_cost.cc",
         "point_to_line_distance_constraint.cc",
@@ -90,6 +92,8 @@ drake_cc_library(
         "gaze_target_constraint.h",
         "kinematic_evaluator_utilities.h",
         "minimum_distance_constraint.h",
+        "minimum_distance_lower_bound_constraint.h",
+        "minimum_distance_upper_bound_constraint.h",
         "orientation_constraint.h",
         "orientation_cost.h",
         "point_to_line_distance_constraint.h",
