+++ multibody/inverse_kinematics/distance_constraint_utilities.cc
@@ -1,6 +1,9 @@
 #include "drake/multibody/inverse_kinematics/distance_constraint_utilities.h"
 
+#include <vector>
+
 #include "drake/math/autodiff_gradient.h"
+#include "drake/multibody/inverse_kinematics/kinematic_evaluator_utilities.h"
 
 namespace drake {
 namespace multibody {
@@ -61,6 +64,87 @@ void CalcDistanceDerivatives(const MultibodyPlant<double>&,
                              double* distance_out) {
   *distance_out = distance_in;
 }
+template <typename T, typename S>
+VectorX<S> Distances(const MultibodyPlant<T>& plant,
+                     systems::Context<T>* context,
+                     const Eigen::Ref<const VectorX<S>>& q,
+                     double influence_distance) {
+  internal::UpdateContextConfiguration(context, plant, q);
+  const auto& query_port = plant.get_geometry_query_input_port();
+  if (!query_port.HasValue(*context)) {
+    throw std::invalid_argument(
+        "MinimumDistanceConstraint: Cannot get a valid geometry::QueryObject. "
+        "Either the plant geometry_query_input_port() is not properly "
+        "connected to the SceneGraph's output port, or the plant_context_ is "
+        "incorrect. Please refer to AddMultibodyPlantSceneGraph on connecting "
+        "MultibodyPlant to SceneGraph.");
+  }
+  const auto& query_object =
+      query_port.template Eval<geometry::QueryObject<T>>(*context);
+
+  const std::vector<geometry::SignedDistancePair<T>> signed_distance_pairs =
+      query_object.ComputeSignedDistancePairwiseClosestPoints(
+          influence_distance);
+  VectorX<S> distances(signed_distance_pairs.size());
+  for (int i = 0; i < static_cast<int>(signed_distance_pairs.size()); ++i) {
+    const geometry::SceneGraphInspector<T>& inspector =
+        query_object.inspector();
+    const geometry::FrameId frame_A_id =
+        inspector.GetFrameId(signed_distance_pairs[i].id_A);
+    const geometry::FrameId frame_B_id =
+        inspector.GetFrameId(signed_distance_pairs[i].id_B);
+    const Frame<T>& frameA = plant.GetBodyFromFrameId(frame_A_id)->body_frame();
+    const Frame<T>& frameB = plant.GetBodyFromFrameId(frame_B_id)->body_frame();
+    internal::CalcDistanceDerivatives(
+        plant, *context, frameA, frameB,
+        // GetPoseInFrame() returns RigidTransform<double> -- we can't
+        // multiply across heterogeneous scalar types; so we cast the double
+        // to T.
+        inspector.GetPoseInFrame(signed_distance_pairs[i].id_A)
+                .template cast<T>() *
+            signed_distance_pairs[i].p_ACa,
+        signed_distance_pairs[i].distance, signed_distance_pairs[i].nhat_BA_W,
+        q, &distances(i));
+  }
+  return distances;
+}
+
+Eigen::VectorXd Distances(
+    const planning::CollisionChecker& collision_checker,
+    planning::CollisionCheckerContext* collision_checker_context,
+    const Eigen::Ref<const Eigen::VectorXd>& x, double influence_distance_val) {
+  return collision_checker
+      .CalcContextRobotClearance(collision_checker_context, x,
+                                 influence_distance_val)
+      .distances();
+}
+
+AutoDiffVecXd Distances(
+    const planning::CollisionChecker& collision_checker,
+    planning::CollisionCheckerContext* collision_checker_context,
+    const Eigen::Ref<const AutoDiffVecXd>& x, double influence_distance_val) {
+  const planning::RobotClearance robot_clearance =
+      collision_checker.CalcContextRobotClearance(collision_checker_context,
+                                                  math::ExtractValue(x),
+                                                  influence_distance_val);
+  return math::InitializeAutoDiff(
+      robot_clearance.distances(),
+      robot_clearance.jacobians() * math::ExtractGradient(x));
+}
+
+// Explicit instantiation
+template VectorX<double> Distances<double, double>(
+    const MultibodyPlant<double>&, systems::Context<double>*,
+    const Eigen::Ref<const VectorX<double>>&, double);
+template VectorX<AutoDiffXd> Distances<double, AutoDiffXd>(
+    const MultibodyPlant<double>&, systems::Context<double>*,
+    const Eigen::Ref<const VectorX<AutoDiffXd>>&, double);
+template VectorX<double> Distances<AutoDiffXd, double>(
+    const MultibodyPlant<AutoDiffXd>&, systems::Context<AutoDiffXd>*,
+    const Eigen::Ref<const VectorX<double>>&, double);
+template VectorX<AutoDiffXd> Distances<AutoDiffXd, AutoDiffXd>(
+    const MultibodyPlant<AutoDiffXd>&, systems::Context<AutoDiffXd>*,
+    const Eigen::Ref<const VectorX<AutoDiffXd>>&, double);
 }  // namespace internal
 }  // namespace multibody
 }  // namespace drake
